requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-ckeditor-5==0.2.10
django-ckeditor==6.7.0
Pillow==10.4.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
